%% File Name: data.mnlfa.Rd
%% File Version: 0.05

\name{data.mnlfa01}
\alias{data.mnlfa01}

\docType{data}
\title{
Example Datasets for \pkg{mnlfa} Package
}
\description{
Example datasets for \pkg{mnlfa} package.
}

\usage{
data(data.mnlfa01)
}



\format{
\itemize{
\item  \code{data.mnlfa01}

  A data frame with 1000 observations for 12 items and 2 covariates.

\code{'data.frame':   1000 obs. of  14 variables:} \cr
\code{ $ female: num  0.5 -0.5 0.5 -0.5 0.5 -0.5 0.5 -0.5 0.5 -0.5 ...} \cr
\code{ $ age   : num  0.79 0.36 0.22 0.79 0.22 -0.34 -0.76 -0.06 0.22 0.65 ...} \cr
\code{ $ I1    : int  1 1 1 1 1 1 1 0 1 1 ...} \cr
\code{ $ I2    : int  0 0 1 1 1 0 1 1 1 1 ...} \cr
\code{ $ I3    : int  1 0 1 0 0 1 0 1 1 1 ...} \cr
\code{ $ I4    : int  1 0 0 1 1 0 1 1 0 1 ...} \cr
\code{ $ I5    : int  1 0 0 0 0 1 1 0 0 1 ...} \cr
\code{ $ I6    : int  1 0 1 1 1 0 1 0 0 1 ...} \cr
\code{ $ I7    : int  1 0 1 1 0 1 1 1 1 1 ...} \cr
\code{ $ I8    : int  0 0 1 0 1 0 1 1 1 1 ...} \cr
\code{ $ I9    : int  1 0 0 1 1 0 0 0 1 0 ...} \cr
\code{ $ I10   : int  0 0 0 0 0 0 0 0 0 1 ...} \cr
\code{ $ I11   : int  0 0 1 0 0 0 0 0 0 1 ...} \cr
\code{ $ I12   : int  0 0 0 0 0 1 1 0 1 0 ...} \cr

 }
}




\keyword{datasets}
