%% File Name: mnlfa-package.Rd
%% File Version: 0.11

\name{mnlfa-package}
% \alias{mnlfa}

\docType{package}

\title{
\packageTitle{mnlfa}
}


\description{
\packageDescription{mnlfa}
}

%\details{
%XXXX
%}

\author{
\packageAuthor{mnlfa}

Maintainer: \packageMaintainer{mnlfa}
}



%\details{
%The \pkg{mnlfa} package enables the estimation of following models:
%}



\references{
Curran, P. J., McGinley, J. S., Bauer, D. J., Hussong, A. M., Burns, A.,
Chassin, L., Sher, K., & Zucker, R. (2014).
A moderated nonlinear factor model for the development of commensurate measures
in integrative data analysis. \emph{Multivariate Behavioral Research, 49}(3),
214-231.
http://dx.doi.org/10.1080/00273171.2014.889594
}



%\seealso{
%For estimating multidimensional models for polytomous item responses
%}


%\examples{
%##   |-----------------------------------------------------------------|
%##   | sirt 0.40-4 (2013-11-26)                                        |
%}

\keyword{package}
